requirements.txt
langchain
langchain-groq
langgraph
yfinance
ta
tavily-python
fastapi
uvicorn
python-dotenv
pandas